Made imports to some modules from some modules
